requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.8.3
python-dotenv==1.0.0
lxml==4.9.3 